refactor: troca do parametro 'makerCode' de query para path na rota 'fipe/veiculos/v1/'
